Actualización de SDK y parámetros requeridos
