+++ openpay-cards/lib/openpay/Openpay/Data/Openpay.php
@@ -1,12 +1,7 @@
 <?php
 
-/**
- * Openpay API v1 Client for PHP (version 2.0.0)
- * 
- * Copyright © Openpay SAPI de C.V. All rights reserved.
- * http://www.openpay.mx/
- * soporte@openpay.mx
- */
+namespace Openpay\Data;
+
 class Openpay
 {
 
@@ -14,17 +9,20 @@ class Openpay
     private static $id = '';
     private static $apiKey = '';
     private static $userAgent = '';
-    private static $country = 'MX';
+    private static $country;
     private static $apiEndpoint = '';
     private static $apiSandboxEndpoint = '';
     private static $sandboxMode = true;
     private static $classification = '';
+    private static $publicIp;
 
-    public function __construct() {
-        
+    public function __construct()
+    {
+
     }
 
-    public static function getInstance($id = '', $apiKey = '', $country = 'MX') {
+    public static function getInstance($id, $apiKey, $country, $publicIp)
+    {
         if ($id != '') {
             self::setId($id);
         }
@@ -35,39 +33,48 @@ class Openpay
             self::setCountry($country);
             self::setEndpointUrl($country);
         }
+        if(!is_null($publicIp)){
+            self::setPublicIp($publicIp);
+        }
         $instance = OpenpayApi::getInstance(null);
         return $instance;
     }
 
-    public static function setUserAgent($userAgent){
+    public static function setUserAgent($userAgent)
+    {
         if ($userAgent != '') {
             self::$userAgent = $userAgent;
         }
     }
 
-    public static function getUserAgent(){
-        $userAgent = self::$userAgent; 
+    public static function getUserAgent()
+    {
+        $userAgent = self::$userAgent;
         return $userAgent;
     }
 
-    public static function setOriginMerchant($classification){
+    public static function setClassificationMerchant($classification)
+    {
         if ($classification != '') {
             self::$classification = $classification;
         }
     }
 
-    public static function getOriginMerchant(){
-        $classification = self::$classification; 
+    public static function getClassificationMerchant()
+    {
+        $classification = self::$classification;
         return $classification;
     }
 
-    public static function setApiKey($key = '') {
+    public static function setApiKey($key = '')
+    {
         if ($key != '') {
             self::$apiKey = $key;
         }
     }
 
-    public static function getApiKey() {
+    public static function getApiKey()
+    {
         $key = self::$apiKey;
         if (!$key) {
             $key = getenv('OPENPAY_API_KEY');
@@ -75,24 +82,28 @@ class Openpay
         return $key;
     }
 
-    public static function setId($id = '') {
+    public static function setId($id = '')
+    {
         if ($id != '') {
             self::$id = $id;
         }
     }
 
-    public static function setCountry($country = ''){
+    public static function setCountry($country = '')
+    {
         if ($country != '') {
             self::$country = $country;
         }
     }
 
-    public static function getCountry(){
+    public static function getCountry()
+    {
         $country = self::$country;
         return $country;
     }
 
-    public static function getId() {
+    public static function getId()
+    {
         $id = self::$id;
         if (!$id) {
             $id = getenv('OPENPAY_MERCHANT_ID');
@@ -100,7 +111,20 @@ class Openpay
         return $id;
     }
 
-    public static function getSandboxMode() {
+    public static function setPublicIp($publicIp = null)
+    {
+        if (!is_null($publicIp)) {
+            self::$publicIp = $publicIp;
+        }
+    }
+
+    public static function getPublicIp() {
+        return self::$publicIp;
+
+    }
+
+    public static function getSandboxMode()
+    {
         $sandbox = self::$sandboxMode;
         if (getenv('OPENPAY_PRODUCTION_MODE')) {
             $sandbox = (strtoupper(getenv('OPENPAY_PRODUCTION_MODE')) == 'FALSE');
@@ -108,11 +132,13 @@ class Openpay
         return $sandbox;
     }
 
-    public static function setSandboxMode($mode) {
+    public static function setSandboxMode($mode)
+    {
         self::$sandboxMode = $mode ? true : false;
     }
 
-    public static function getProductionMode() {
+    public static function getProductionMode()
+    {
         $sandbox = self::$sandboxMode;
         if (getenv('OPENPAY_PRODUCTION_MODE')) {
             $sandbox = (strtoupper(getenv('OPENPAY_PRODUCTION_MODE')) == 'FALSE');
@@ -120,64 +146,33 @@ class Openpay
         return !$sandbox;
     }
 
-    public static function setProductionMode($mode) {
+    public static function setProductionMode($mode)
+    {
         self::$sandboxMode = $mode ? false : true;
     }
 
-    public static function setEndpointUrl($country){
-        if($country == 'MX'){
-            if(self::getOriginMerchant() != 'eglobal'){
+    public static function setEndpointUrl($country)
+    {
+        if ($country == 'MX') {
+            if (self::getClassificationMerchant() != 'eglobal') {
                 self::$apiEndpoint = 'https://api.openpay.mx/v1';
                 self::$apiSandboxEndpoint = 'https://sandbox-api.openpay.mx/v1';
-            }else{
+            } else {
                 self::$apiEndpoint = 'https://api.ecommercebbva.com/v1';
                 self::$apiSandboxEndpoint = 'https://sand-api.ecommercebbva.com/v1';
             }
-        }elseif($country == 'CO'){
+        } elseif ($country == 'CO') {
             self::$apiEndpoint = 'https://api.openpay.co/v1';
             self::$apiSandboxEndpoint = 'https://sandbox-api.openpay.co/v1';
-        }elseif($country == 'PE'){
+        } elseif ($country == 'PE') {
             self::$apiEndpoint = 'https://api.openpay.pe/v1';
             self::$apiSandboxEndpoint = 'https://sandbox-api.openpay.pe/v1';
         }
     }
-    public static function getEndpointUrl() {
+
+    public static function getEndpointUrl()
+    {
         return (self::getSandboxMode() ? self::$apiSandboxEndpoint : self::$apiEndpoint);
     }
 
 }
-
-// ----------------------------------------------------------------------------
-class OpenpayApi extends OpenpayApiResourceBase
-{
-
-    protected $derivedResources = array(
-        'Bine' => array(),
-        'Customer' => array(),
-        'Card' => array(),
-        'Charge' => array(),
-        'Pse' => array(),
-        'Payout' => array(),
-        'Fee' => array(),
-        'Plan' => array(),
-        'Webhook' => array(),
-        'Token' => array());
-
-    public static function getInstance($r, $p = null) {
-        $resourceName = get_class();
-        return parent::getInstance($resourceName);
-    }
-    public function getMerchantInfo(){
-        return parent::getMerchantInfo();
-    }
-    protected function getResourceUrlName($p = true) {
-        return '';
-    }
-
-    public function getFullURL() {
-        return $this->getUrl();
-    }
-
-}
-
-?>
