+++ openpay-cards/lib/openpay/Openpay/Data/OpenpayApiAuthError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Openpay\Data;
+
+class OpenpayApiAuthError extends OpenpayApiError
+{
+
+}
