+++ openpay-cards/lib/openpay/Openpay/Data/OpenpayApiConnectionError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Openpay\Data;
+
+class OpenpayApiConnectionError extends OpenpayApiError
+{
+
+}
