+++ openpay-cards/lib/openpay/Openpay/Data/OpenpayApiConsole.php
@@ -1,13 +1,8 @@
 <?php
 
-/**
- * Openpay API v1 Client for PHP (version 2.0.0)
- * 
- * Copyright © Openpay SAPI de C.V. All rights reserved.
- * http://www.openpay.mx/
- * soporte@openpay.mx
- */
-class OpenpayConsole {
+namespace Openpay\Data;
+
+class OpenpayApiConsole {
 
     const CONSOLE_NONE = 0;
     const CONSOLE_TRACE = 2;
