+++ openpay-cards/lib/openpay/Openpay/Data/OpenpayApiDerivedResource.php
@@ -1,18 +1,13 @@
 <?php
 
-/**
- * Openpay API v1 Client for PHP (version 2.0.0)
- * 
- * Copyright © Openpay SAPI de C.V. All rights reserved.
- * http://www.openpay.mx/
- * soporte@openpay.mx
- */
+namespace Openpay\Data;
+
 class OpenpayApiDerivedResource extends OpenpayApiResourceBase {
 
     private $cacheList = array();
 
     protected static function getInstance($resourceName, $p = null) {
-        if (class_exists($resourceName . 'List', false)) {
+        if (class_exists($resourceName . 'List')) {
             $resource = $resourceName . 'List';
             return new $resource($resourceName);
         }
@@ -57,7 +52,7 @@ class OpenpayApiDerivedResource extends OpenpayApiResourceBase {
 
 
     public function add($params) {
-        OpenpayConsole::trace('OpenpayApiDerivedResource @add');
+        OpenpayApiConsole::trace('OpenpayApiDerivedResource @add');
 
         // TODO: validate call when the parent has not a valid ID
         $resource = parent::_create($this->resourceName, $params, array('parent' => $this));
@@ -66,7 +61,7 @@ class OpenpayApiDerivedResource extends OpenpayApiResourceBase {
     }
 
     public function get($id) {
-        OpenpayConsole::trace('OpenpayApiDerivedResource @get');
+        OpenpayApiConsole::trace('OpenpayApiDerivedResource @get');
 
         if ($this->isResourceListed($id)) {
             return $this->getResource($id);
@@ -77,7 +72,7 @@ class OpenpayApiDerivedResource extends OpenpayApiResourceBase {
     }
 
     public function getList($params) {
-        OpenpayConsole::trace('OpenpayApiDerivedResource @find');
+        OpenpayApiConsole::trace('OpenpayApiDerivedResource @find');
 
         $list = parent::_find($this->resourceName, $params, array('parent' => $this));
         foreach ($list as $resource) {
