+++ openpay-cards/lib/openpay/Openpay/Data/OpenpayApiError.php
@@ -1,12 +1,9 @@
 <?php
 
-/**
- * Openpay API v1 Client for PHP (version 2.0.0)
- * 
- * Copyright © Openpay SAPI de C.V. All rights reserved.
- * http://www.openpay.mx/
- * soporte@openpay.mx
- */
+namespace Openpay\Data;
+
+use Exception;
+
 class OpenpayApiError extends Exception
 {
 
@@ -17,7 +14,7 @@ class OpenpayApiError extends Exception
     protected $request_id;
     protected $fraud_rules;
 
-    public function __construct($message = null, $error_code = null, $category = null, $request_id = null, $http_code = null, $fraud_rules = null) {
+    public function __construct($message = null, $error_code = 0, $category = null, $request_id = null, $http_code = null, $fraud_rules = null) {
         parent::__construct($message, $error_code);
         $this->description = $message;
         $this->error_code = isset($error_code) ? $error_code : 0;
@@ -51,28 +48,4 @@ class OpenpayApiError extends Exception
         return $this->fraud_rules;
     }
 
-}
-
-// Authentication related Errors
-class OpenpayApiAuthError extends OpenpayApiError
-{
-    
-}
-
-// Request related Error
-class OpenpayApiRequestError extends OpenpayApiError
-{
-    
-}
-
-// Transaction related Errors
-class OpenpayApiTransactionError extends OpenpayApiError
-{
-    
-}
-
-// Connection related Errors
-class OpenpayApiConnectionError extends OpenpayApiError
-{
-    
 }
