+++ openpay-cards/lib/openpay/Openpay/Data/OpenpayApiRequestError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Openpay\Data;
+
+class OpenpayApiRequestError extends OpenpayApiError
+{
+
+}
