+++ openpay-cards/lib/openpay/Openpay/Data/OpenpayApiTransactionError.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Openpay\Data;
+
+class OpenpayApiTransactionError extends OpenpayApiError
+{
+
+}
