+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayBankAccountList.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiDerivedResource;
+
+class OpenpayBankAccountList extends OpenpayApiDerivedResource
+{
+}
