+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayBine.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiResourceBase;
+
+class OpenpayBine extends OpenpayApiResourceBase {
+
+}
