+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayCapture.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiResourceBase;
+
+class OpenpayCapture extends OpenpayApiResourceBase {
+	protected function getResourceUrlName($p = true){
+		return parent::getResourceUrlName(false);
+	}
+}
+?>
