+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayCard.php
@@ -1,12 +1,17 @@
 <?php
 
 /**
- * Openpay API v1 Client for PHP (version 1.0.0)
- * 
+ * Openpay API v1 Client for PHP (version 2.2.*)
+ *
  * Copyright © Openpay SAPI de C.V. All rights reserved.
  * http://www.openpay.mx/
  * soporte@openpay.mx
  */
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiResourceBase;
+use Openpay\Data\OpenpayApiDerivedResource;
+
 class OpenpayCard extends OpenpayApiResourceBase
 {
 
@@ -30,9 +35,5 @@ class OpenpayCard extends OpenpayApiResourceBase
 }
 
 // ----------------------------------------------------------------------------
-class OpenpayCardList extends OpenpayApiDerivedResource
-{
-    
-}
 
 ?>
