+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayCardList.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiDerivedResource;
+
+class OpenpayCardList extends OpenpayApiDerivedResource
+{
+
+}
