+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayCustomer.php
@@ -1,12 +1,10 @@
 <?php
 
-/**
- * Openpay API v1 Client for PHP (version 1.0.0)
- * 
- * Copyright © Openpay SAPI de C.V. All rights reserved.
- * http://www.openpay.mx/
- * soporte@openpay.mx
- */
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiResourceBase;
+use Openpay\Data\OpenpayApiDerivedResource;
+
 class OpenpayCustomer extends OpenpayApiResourceBase {
 
     protected $status;
@@ -32,9 +30,5 @@ class OpenpayCustomer extends OpenpayApiResourceBase {
 
 }
 
-// ----------------------------------------------------------------------------
-class OpenpayCustomerList extends OpenpayApiDerivedResource {
-    
-}
 
 ?>
