+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayCustomerList.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiDerivedResource;
+
+class OpenpayCustomerList extends OpenpayApiDerivedResource
+{
+
+}
