+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayFee.php
@@ -1,12 +1,10 @@
 <?php
 
-/**
- * Openpay API v1 Client for PHP (version 1.0.0)
- * 
- * Copyright © Openpay SAPI de C.V. All rights reserved.
- * http://www.openpay.mx/
- * soporte@openpay.mx
- */
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiResourceBase;
+use Openpay\Data\OpenpayApiDerivedResource;
+
 class OpenpayFee extends OpenpayApiResourceBase
 {
 
@@ -17,7 +15,7 @@ class OpenpayFee extends OpenpayApiResourceBase
     protected $status;
     protected $transaction_type;
     protected $error_message;
-    protected $method;    
+    protected $method;
     protected $derivedResources = array('Refund' => null);
 
     public function refund($params) {
@@ -25,16 +23,6 @@ class OpenpayFee extends OpenpayApiResourceBase
         if ($resource) {
             return parent::_create($resource->resourceName, $params, array('parent' => $this));
         }
-    }    
-
-}
-
-// ----------------------------------------------------------------------------
-class OpenpayFeeList extends OpenpayApiDerivedResource
-{
-
-    public function create($params) {
-        return $this->add($params);
     }
 
 }
