+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayPayoutList.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiDerivedResource;
+
+class OpenpayPayoutList extends OpenpayApiDerivedResource
+{
+    public function create($params)
+    {
+        return $this->add($params);
+    }
+}
