+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayPlanList.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiDerivedResource;
+
+class OpenpayPlanList extends OpenpayApiDerivedResource
+{
+}
