+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayPseList.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiDerivedResource;
+
+class OpenpayPseList extends OpenpayApiDerivedResource
+{
+
+    public function create($params)
+    {
+        return $this->add($params);
+    }
+
+}
