+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayRefund.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiResourceBase;
+
+class OpenpayRefund extends OpenpayApiResourceBase {
+	protected function getResourceUrlName($p = true){
+		return parent::getResourceUrlName(false);
+	}
+}
+?>
