+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayToken.php
@@ -1,12 +1,9 @@
 <?php
 
-/**
- * Openpay API v1 Client for PHP (version 1.0.0)
- * 
- * Copyright © Openpay SAPI de C.V. All rights reserved.
- * http://www.openpay.mx/
- * soporte@openpay.mx
- */
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiResourceBase;
+
 class OpenpayToken extends OpenpayApiResourceBase
 {
     protected $card;
