+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayTransfer.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Openpay\Resources;
+
+
+use Openpay\Data\OpenpayApiResourceBase;
+use Openpay\Data\OpenpayApiDerivedResource;
+
+class OpenpayTransfer extends OpenpayApiResourceBase {
+	protected $authorization;
+	protected $creation_date;
+	protected $currency;
+	protected $operation_type;
+	protected $status;
+	protected $transaction_type;
+	protected $error_message;
+	protected $method;
+}
+?>
