+++ openpay-cards/lib/openpay/Openpay/Resources/OpenpayWebhook.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Openpay\Resources;
+
+use Openpay\Data\OpenpayApiResourceBase;
+
+class OpenpayWebhook extends OpenpayApiResourceBase
+{
+
+    protected $url;
+    protected $event_types;
+
+    public function save()
+    {
+        return $this->_update();
+    }
+
+    public function delete()
+    {
+        $this->_delete();
+    }
+
+}
+
+?>
