--- openpay-cards/lib/openpay/resources/OpenpayBine.php
@@ -1,14 +1,0 @@
-<?php
-
-/**
- * Openpay API v1 Client for PHP (version 2.0.1)
- * 
- * Copyright © Openpay SAPI de C.V. All rights reserved.
- * http://www.openpay.mx/
- * soporte@openpay.mx
- */
-class OpenpayBine extends OpenpayApiResourceBase {
-
-}
-
-?>
