--- openpay-cards/lib/openpay/resources/OpenpayRefund.php
@@ -1,15 +1,0 @@
-<?php 
-/**
- * Openpay API v1 Client for PHP (version 1.0.0)
- * 
- * Copyright © Openpay SAPI de C.V. All rights reserved.
- * http://www.openpay.mx/
- * soporte@openpay.mx
- */
-
-class OpenpayRefund extends OpenpayApiResourceBase {
-	protected function getResourceUrlName($p = true){
-		return parent::getResourceUrlName(false);
-	}
-}
-?>
